Update background task
Old method used a global list for all post from all tasks. This approach was not a smart one.
Changed it so that each task will work with it's own list and not worry about anyone else.
Logic remains the same, posts are added to the front and popped from the end.

Closes #13