Better update to background task
We don't care about first, last, insert here or insert there. The only thing we care about is NEW.
If a new post is found, post it, that's it!

New method will create two list, one initial one outside the loop and another inside the loop. If we find anything that is not in the outside one, then that is a new submission. Post that one.

This one completely overwrites the previous commit because this has better logic without any of the issues the previous [commit](798fac78af05c07bdc5728b603945ff83c764e03) had anyways.

Closes #13