Done and Allow All button functional
Done and Allow All button correctly finds the buttons that were clicked or all of them.
Fixed issue with some subreddits that send a 403 Forbidden attribute. An empty list is used in these cases.
Done button is disabled until something is clicked.
However, if a flair is added and then removed, the Done button will be incorrectly enabled.